==> Flight_analysis/notebooks/01_Download_data.ipynb <==
# Notebook description

The role of this notebook is to retrieve data and save it to the local disk.

The API prepared by CodersLab was used https://api-datalab.coderslab.com/api/v2. Including documentation: [click](https://api-datalab.coderslab.com/v2/docs/).

In order to retrieve the relevant data, a list from the `airports.csv` file was used.


 ## Notepad configuration

import pandas as pd 
import requests 
import os
from dotenv import  load_dotenv
from datetime import datetime
from dateutil.relativedelta import relativedelta

load_dotenv("D:\Python\devcontainer\.env") #loading API key from .env file

 API configuration

APIKEY = os.getenv("CODERS_PORT_APIKEY")

 Loading the file `airports.csv`

df_airports = pd.read_csv(r'D:\Python\Portfolio_lab\data\airports.csv')

 # Download `Airport`

df_raw_airports = []

for id in df_airports['origin_airport_id']:
    response = requests.get(f'https://api-datalab.coderslab.com/api/v2/airport/{id}', headers={'Authorization': f'{APIKEY}'})
    if response.status_code == 200:
        data = response.json()
        df_raw_airports.append(data)
    else:
        pass

airport_df = pd.DataFrame(df_raw_airports)

airport_df.head()

Save `airport_df` to `airport_list.csv` file.

airport_df.to_csv('airport_list.csv', index=False)

# Download  `Weather`

start_date = datetime(2019, 1, 1)
end_date = datetime(2020, 3, 31)

dates = []


date = start_date
while date <= end_date:
    dates.append(date.strftime('%Y-%m'))
    date += relativedelta(months=1)


df_raw_weather = []
timer = 0


for date in dates:
    response = requests.get(f'https://api-datalab.coderslab.com/api/v2/airportWeather?date={date}', headers={'Authorization': f'{APIKEY}'})
    data = response.json()
    df_raw_weather.append(data)

airport_weather_df = pd.DataFrame([dictt for list in df_raw_weather for dictt in list])

airport_weather_df.head()

Save `weather_df` to `airport_weather.csv` file.

airport_weather_df.to_csv('airport_weather.csv', index=False)

 # Download `Aircraft`

response = requests.get(f'https://api-datalab.coderslab.com/api/v2/aircraft', headers={'Authorization': f'{APIKEY}'})
data = response.json()

aircraft_df = pd.DataFrame(data)

aircraft_df.head()

Save `aircraft_df` to `aircraft.csv` file.

aircraft_df.to_csv('aircraft.csv', index=False)    

 # Download `Flight`

start_date = datetime(2019, 1, 1)
end_date = datetime(2020, 3, 31)

dates = []


date = start_date
while date <= end_date:
    dates.append(date.strftime('%Y-%m'))
    date += relativedelta(months=1)

df_airports.head()

flight_data = []

for airport_id in df_airports['origin_airport_id']:
    for date in dates:
        response = requests.get(f'https://api-datalab.coderslab.com/api/v2/flight?airportId={airport_id}&date={date}', 
                        headers={'Authorization': f'{APIKEY}'})
        if response.status_code == 200:
            f_data = response.json()
            flight_data.append(f_data)
        elif response.status_code == 204:
            pass

flight_df = pd.DataFrame([dictt for list in flight_data for dictt in list])

flight_df.shape

flight_df.head()

Save `flight_df` to `flight.csv` file.

flight_df.to_csv('flight2.csv', index=False)
