==> Flight_analysis/notebooks/02_Creating_database.ipynb <==
# Notebook description

The role of this notebook is to create a database with a table structure. 

For this task, `Postgres' was used. 

## Notepad configuration

import psycopg2
import os
from dotenv import  load_dotenv

load_dotenv("D:\Python\devcontainer\.env") #loading API key from .env file

 ## Database connection

username = 'postgres'
password = os.getenv("POSTGRES_PASSWORD")

host = 'localhost'
database = 'postgres'
port = None

con = psycopg2.connect(host='localhost',user='postgres', password = os.getenv("POSTGRES_PASSWORD"))
cur = con.cursor()

 ## Load the file `database_schema.sql`.

sql_file_base = 'Projects\Flight_analysis\sql\database_schema.sql'

with open(sql_file_base, 'r') as file:
    base = file.read()

Split the contents of the `database_schema.sql` file into smaller queries using `;`.

base_split = base.split(';')

Execute each of the queries to initialize the database structure

for query in base_split:
    cur.execute(query)

Approval of all operations on the base.

con.commit()
