==> Flight_analysis/notebooks/03_Uploading_Data.ipynb <==
 # Notebook description

 The role of this notebook is to adjust the data structure from the source files to a format compatible with `Postgres`, and then upload them to the server.

## Notepad configuration

from sqlalchemy import create_engine
from sqlalchemy.engine import URL
import pandas as pd 
import os
from dotenv import  load_dotenv

load_dotenv("D:\Python\devcontainer\.env") #loading postgre password from .env file

Create url and engine

url = URL.create(
    "postgresql+psycopg2",
    username='postgres',
    password=os.getenv("POSTGRES_PASSWORD"),  
    host='localhost',
    database='postgres',
)
engine = create_engine(url)

Loading frames into the workspace

def load_raw_data(file_name):
    file_name.columns = [column.lower() for column in file_name.columns]
    return file_name 

Uploading individual files to frames

df_aircraft = pd.read_csv('aircraft.csv')

df_airport = pd.read_csv('airport_list.csv')

df_weather = pd.read_csv('airport_weather.csv')

df_flight = pd.read_csv('flight.csv')

load_raw_data(df_aircraft)

df_aircraft.sample()

load_raw_data(df_airport)

df_airport.sample()

load_raw_data(df_weather)

df_weather.sample()

load_raw_data(df_flight)

df_flight.sample()

### Exporting data to a database

def export_table_to_df(table_name, df):
    df.to_sql(name=table_name, con=engine, if_exists='append', index=False)

 ## Uploading data

 ### Uploading `aircraft_df` to the `aircraft` table.

df_aircraft.to_sql(name='aircraft', con=engine, if_exists='append', index=False)

 ### Uploading `airport_weather_df` to the `airport_weather` table.

df_weather.to_sql(name='airport_weather', con=engine, if_exists='append', index=False)

 ### Uploading `flight_df` to the `flight` table.

df_flight.to_sql(name='flight', con=engine, if_exists='append', index=False)

 ### Uploading `airport_list_df` to the `airport_list` table.

df_airport.to_sql(name='airport_list', con=engine, if_exists='append', index=False)
