==> Flight_analysis/notebooks/04_Data_Analysis_02.ipynb <==
 # Notebook description

 This notebook is a counter-analysis of flight data and delays.

## Notepad configuration

from sqlalchemy import create_engine
from sqlalchemy.engine import URL
import pandas as pd 
import plotly.express as px
import os
from dotenv import  load_dotenv

load_dotenv("D:\Python\devcontainer\.env") #loading postgre password from .env file

Create url and engine

url = URL.create(
    "postgresql+psycopg2",
    username='postgres',
    password=os.getenv("POSTGRES_PASSWORD"),  
    host='localhost',
    database='postgres',
)
engine = create_engine(url)

Loading an SQL database using Pandas

def read_sql_table(table_name):
    df = pd.read_sql(sql=table_name, con=engine)
    return df

flight_df = pd.read_csv(r'flight_df_01.csv') #loading file flight_df_01.csv created at the end of the notebook 04_Data_Analysis_01.ipynb 

 ## Loading the `aircraft` table


aircraft_df = read_sql_table(table_name='aircraft')

Delete `number_of_seats`, `id` and duplicate columns from the `aircraft_df` frame.

aircraft_df = aircraft_df.drop(columns=['id', 'number_of_seats'])

aircraft_df = aircraft_df.drop_duplicates()

Check if there is more than one year of production for a given `tail_num`.

aircraft_df.duplicated('tail_num').any()

aircraft_df_duplicated = aircraft_df.loc[aircraft_df.duplicated(subset='tail_num')]

aircraft_df_duplicated.head()

 ## Modification `aircraft_df`.


For duplicated `tail_num`, `manufacture_year` has been set as the highest

aircraft_df['manufacture_year'] = aircraft_df.groupby('tail_num')['manufacture_year'].transform('max')

 ## Merge `aircraft_df` and `flight_df`.

Merged tables saved to `tmp_flight_df`.

tmp_flight_df = flight_df.merge(aircraft_df, how='left', on='tail_num')

Check for duplicates

tmp_flight_df.duplicated().any()

tmp_flight_df.shape

tmp_flight_df = tmp_flight_df.drop_duplicates()

tmp_flight_df.shape

Overwrite the `flight_df` table using `tmp_flight_df`.

flight_df = tmp_flight_df

 ## Delays vs. year of aircraft production
Determination of the relation of the year of production to the frequency of delays

delays_by_manufacture_year_df = flight_df.groupby('manufacture_year')['is_delayed'].mean().reset_index()

Chart Manufacture Year vs. Delay

fig = px.scatter(delays_by_manufacture_year_df, 
                    x = 'manufacture_year',
                    y = 'is_delayed',
                    title= 'Manufacture Year vs. Delay'
                   )

fig.update_xaxes(title_text='Manufacture Year')

fig.update_yaxes(title_text='Delayed Flights [%]')

fig.show()

Chart modification: `manufacture_year` that have flown more than 10,000 flights.

delays_by_manufacture_year_df = flight_df.groupby('manufacture_year')['is_delayed'].agg(['mean', 'count']).reset_index()

fig = px.scatter(delays_by_manufacture_year_df[delays_by_manufacture_year_df['count'] > 10000], 
                    x = 'manufacture_year',
                    y = 'mean',
                    title= 'Manufacture Year vs. Delay'
                   )

fig.update_xaxes(title_text='Manufacture Year')

fig.update_yaxes(title_text='Delayed Flights [%]')

fig.show()

Adding a `manufacture_year_agg` column to the `flight_df` frame.

flight_df['manufacture_year_agg'] = pd.cut(flight_df['manufacture_year'], bins = range(int(min(flight_df['manufacture_year'])), int(max(flight_df['manufacture_year'])) + 3, 3), right = True)

delays_by_manufacture_year_df.sample()

Creation of the variable `flight_delays_by_manufacture_year_agg_df`.

flight_delays_by_manufacture_year_agg_df = flight_df.groupby('manufacture_year_agg')['is_delayed'].agg(['mean', 'count']).reset_index()

flight_delays_by_manufacture_year_agg_df['manufacture_year_agg'] = flight_delays_by_manufacture_year_agg_df['manufacture_year_agg'].astype(str)

Chart based on data in `flight_delays_by_manufacture_year_agg_df`.

fig = px.scatter(flight_delays_by_manufacture_year_agg_df, 
                    x = 'manufacture_year_agg',
                    y = 'mean',
                    title= 'Manufacture Year vs. Delay'
                   )

fig.update_xaxes(title_text='Manufacture Year')

fig.update_yaxes(title_text='Delayed Flights [%]')

fig.show()

TOP 5 production years with the most flights

top_manufactured_df = delays_by_manufacture_year_df.sort_values('count', ascending=False).head()

top_manufactured_df.head()

Saving the frame to .cvs

flight_df.to_csv('flight_df_02.csv', index=False)
