==> Flight_analysis/notebooks/04_Data_Analysis_03.ipynb <==
 # Notebook description

 This notebook is a counter-analysis of flight data and delays.

Notepad configuration

from sqlalchemy import create_engine
from sqlalchemy.engine import URL
import pandas as pd 
import plotly.express as px
import os
from dotenv import  load_dotenv

load_dotenv("D:\Python\devcontainer\.env") #loading postgres password from .env file

Create url and engine

url = URL.create(
    "postgresql+psycopg2",
    username='postgres',
    password=os.getenv("POSTGRES_PASSWORD"),  
    host='localhost',
    database='postgres',
)
engine = create_engine(url)

Loading an SQL database using Pandas

def read_sql_table(table_name):
    df = pd.read_sql(sql=table_name, con=engine)
    return df

 Loading data frame `flight_df`

flight_df = pd.read_csv(r'flight_df_02.csv') #loading file flight_df_02.csv created at the end of the notebook 04_Data_Analysis_02.ipynb 

 #  `airport_list`

Loading data frame `airport_list_df`

airport_list_df = read_sql_table(table_name='airport_list')

Checking duplicates for a column `origin_airport_id`

airport_list_df.origin_airport_id.duplicated().any()

Merge `flight_df` and `airport_list_df` using `origin_airport_id`

flight_df = flight_df.merge(airport_list_df[['origin_airport_id', 'origin_city_name']], how='left', on='origin_airport_id')

Merge `flight_df` and `airport_list_df` using `destination_airport_id`

flight_df = flight_df.merge(airport_list_df[['origin_airport_id', 'origin_city_name']], how='left', left_on='dest_airport_id', right_on='origin_airport_id')

flight_df = flight_df.rename(columns={'origin_city_name_x': 'origin_city_name', 'origin_city_name_y': 'destination_city_name', 'origin_airport_id_x': 'origin_airport_id'}).drop(columns='origin_airport_id_y')


 ## Analysis by airports and routes

Determine the airport from which the most planes departed.

flight_df.sample()

top_airports_origin_df = flight_df.groupby('origin_city_name').agg('count').sort_values('id', ascending=False).reset_index()

top_airports_origin_df.head()

Determine the airport to which the most planes arrived

top_airports_destination_df = flight_df.groupby('destination_city_name').agg('count').sort_values('id', ascending=False).reset_index()

 # Adding weather data


 Loading data frame `airport_weather`

airport_weather_df = read_sql_table(table_name='airport_weather')

Deleting redundant columns

airport_weather_df.columns

airport_weather_df = airport_weather_df.drop(columns=['id','pgtm',
       'tavg', 'tmin', 'wdf2', 'wdf5', 'wsf2', 'wsf5', 'wt01', 'wt02',
       'wt03', 'wt04', 'wt05', 'wt06', 'wt07', 'wt08', 'wt09', 'wesd', 'wt10',
       'psun', 'tsun', 'sn32', 'sx32', 'tobs', 'wt11', 'wt18'])

Merge `airport_list_df` and `airport_weather_df` using updating `airport_weather_df`.

airport_weather_df = airport_weather_df.merge(airport_list_df[['origin_airport_id', 'name']], on='name', how='inner')

 ## Merge `airport_weather_df` and `flight_df`.

Change the `date` column type to `DATETIME` in the `airport_weather_df` frame.

airport_weather_df['date'] = pd.to_datetime(airport_weather_df['date'])

airport_weather_df.info()

Create a date` column in the `flight_df` frame. 

flight_df = flight_df.rename(columns={'day_of_month': 'day'})

flight_df['date'] = pd.to_datetime(flight_df[['year', 'month', 'day']])

flight_df['date'] = pd.to_datetime(flight_df['date']).dt.strftime('%Y-%m-%d')

flight_df['date'] = pd.to_datetime(flight_df['date'])

Merge `airport_weather_df` and `flight_df` based on: `origin_airport_id`, `date`.

flight_df = flight_df.merge(airport_weather_df, on=['origin_airport_id', 'date'], how='left')

flight_df.sample()

## Analysis of the effect of maximum temperature on delays

Analysis for the column `tmax`.

flight_df.sample()

Outliers

flight_df.tmax.describe(percentiles=[0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99]).round(2)

flight_df[flight_df['tmax'] > 114].origin_city_name.unique()

 ### Comment
temperatures around 115 °F occur in Phoenix, AZ naturally during the period defined by the data, i.e. from June to September (the highest recorded 122 °F), so it is not dismissed as an outlier

source:
- https://www.currentresults.com/Weather/Arizona/Places/phoenix-temperatures-by-month-average.php
- https://www.azfamily.com/page/what-they-dont-tell-you-about-arizonas-temperature-extremes/

Removal of Null in tmax

flight_df.tmax.isnull().any()

tmax_clean_flight_df = flight_df.dropna(subset=['tmax'])

tmax_clean_flight_df.tmax.isnull().any()

tmax_clean_flight_df.sample()

agg_tmax_clean_flight_df = tmax_clean_flight_df.groupby('tmax')['is_delayed'].agg(['count', 'mean']).reset_index()

agg_tmax_clean_flight_df.sort_values('tmax')

agg_tmax_clean_flight_df.sample()

Graph of the effect of temperature on delay

fig = px.scatter(agg_tmax_clean_flight_df, 
                    x = 'tmax',
                    y = 'mean',
                    title= 'Temperature vs. Delay'
                   )

fig.update_xaxes(title_text='Temperature °F')

fig.update_yaxes(title_text='Delayed Flights [%]')

fig.show()

## Comments on the results

Temperature affects the percentage of delays up to 40°F, or about 4.4°C, which corresponds to reality, such as the need to de-ice the aircraft, which leads to delays, and the constant maintenance of the runway in conditions for takeoff and landing.
