==> Flight_analysis/notebooks/05_Creating_a_reporting_scheme.ipynb <==
# Notebook description

Create a `reporting` schema based on the queries in the `reporting.sql` file. To create a report from previously obtained data.

 ## Database update

Notepad configuration

import psycopg2
import os
from dotenv import  load_dotenv

load_dotenv("D:\Python\devcontainer\.env") #loading postgres password from .env file

Create url and engine

con = psycopg2.connect(host='localhost',
                       user='postgres', 
                       password=os.getenv("POSTGRES_PASSWORD"))
cur = con.cursor()

Loading `reporting.sql` file

sql_file_base = r'Projects\Flight_analysis\sql\reporting.sql'

Split the contents of the `reporting.sql` file into smaller queries using `;`.

with open(sql_file_base, 'r') as file:
    base = file.read()

base_split = base.split(';')

Execution of each of the queries

for query in base_split:
    cur.execute(query)

 Approval of all operations on the base

con.commit()

Closing the connection

con.close()
